Version 0.1.2?
In this version I have added in the "{ws}" declaration to replace all of the "\s" so everything is more consistent with how Flex works.  Removed all of the old code to handle INCLUDES and replaced with code found on http://flex.sourceforge.net/manual/Multiple-Input-Buffers.html.  Mainly because it is simple and it works.