Version 0.1.2
Changes for the 0.1.2 version in the cmds.php file.  The changes include swapping out the "\s" with the "{ws}".  The cmds.php script is comming to the end of is usefulness.  This is because there really aren't any new commands to put in.  So most of the modifications will now be in the bas.l file as well as the bas.y file when I start working on the parser.