Version 0.1.3
Changed out all of the commands with more basic style (rather than with a lot of [aA] kind of stuff in it because I've already turned on the insensitive flag.  I also fixed the $include part yet again.  This time I put the REM and single quote (') parts in.  Also changed out the standard REM and single quote (') commands so they work properly.  The BAS.L file looks a bit nicer now.