Version 0.2
In this version I have added in a simple string finder so basic strings like "This is a test" can be found.  It will also handle things like "This is my quote \"Hi ya!\"" correctly.  Presently it doesn't allow you to use single quotes (') for strings.  I need to look at the QB64 manual to see if those are allowed before I put them in.  I also implemented a START_STACK.  What I mean by that is, if you have multiple start conditions (like the INCL or include handler and now the STRING handler) you have to keep track of what was last invoked. This is because Flex (and Lex both) only have a BEGIN command and no END command (which seems a bit weird to me - but anyway).  So you have to create a stack that you can push and pull the various start states so you don't get into trouble (which I did). This makes it so if there are includes inside of includes with strings inside of each include - the parser will keep everything in its proper order and pop back and forth between whatever state it needs to be in up to the maximum depth allowed.  The maximum depth allowed for the stack has to be at least two times as big as the include stack so it can keep track of everything.  In reality though - I am making it three times as big just in case.  3,000 INTs are not going to kill us.