Version 0.21
Added in code to bas.l to handle yylval.  Set it to be a plain character array
at this point.

Added in code to the main program to display the returning yylval values.

Added in code to bas.l to make all of the expressions handle the storage of
the incoming text into the yylval variable.